requirements.txt
annotated-types==0.7.0 ; python_version >= "3.12" and python_version < "4.0"
anyio==4.6.0 ; python_version >= "3.12" and python_version < "4"
brotli==1.1.0 ; python_version >= "3.12" and python_version < "4.0" and implementation_name == "cpython"
brotlicffi==1.1.0.0 ; python_version >= "3.12" and python_version < "4.0" and implementation_name != "cpython"
cachetools==5.5.0 ; python_version >= "3.12" and python_version < "4.0"
certifi==2024.8.30 ; python_version >= "3.12" and python_version < "4.0"
cffi==1.17.1 ; python_version >= "3.12" and python_version < "4.0" and implementation_name != "cpython"
charset-normalizer==3.4.0 ; python_version >= "3.12" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "4.0" and platform_system == "Windows"
google-ai-generativelanguage==0.6.10 ; python_version >= "3.12" and python_version < "4.0"
google-api-core==2.21.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-core[grpc]==2.21.0 ; python_version >= "3.12" and python_version < "4.0"
google-api-python-client==2.149.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth-httplib2==0.2.0 ; python_version >= "3.12" and python_version < "4.0"
google-auth==2.35.0 ; python_version >= "3.12" and python_version < "4.0"
google-generativeai==0.8.3 ; python_version >= "3.12" and python_version < "4.0"
googleapis-common-protos==1.65.0 ; python_version >= "3.12" and python_version < "4.0"
grpcio-status==1.66.2 ; python_version >= "3.12" and python_version < "4.0"
grpcio==1.66.2 ; python_version >= "3.12" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.12" and python_version < "4"
httpcore==1.0.6 ; python_version >= "3.12" and python_version < "4"
httplib2==0.22.0 ; python_version >= "3.12" and python_version < "4.0"
httpx==0.27.2 ; python_version >= "3.12" and python_version < "4"
idna==3.10 ; python_version >= "3.12" and python_version < "4.0"
mutagen==1.47.0 ; python_version >= "3.12" and python_version < "4.0"
notion-client==2.2.1 ; python_version >= "3.12" and python_version < "4"
numpy==2.1.2 ; python_version >= "3.12" and python_version < "4.0"
opencv-python-headless==4.10.0.84 ; python_version >= "3.12" and python_version < "4.0"
proto-plus==1.24.0 ; python_version >= "3.12" and python_version < "4.0"
protobuf==5.28.2 ; python_version >= "3.12" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.12" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.12" and python_version < "4.0"
pycparser==2.22 ; python_version >= "3.12" and python_version < "4.0" and implementation_name != "cpython"
pycryptodomex==3.21.0 ; python_version >= "3.12" and python_version < "4.0"
pydantic-core==2.23.4 ; python_version >= "3.12" and python_version < "4.0"
pydantic==2.9.2 ; python_version >= "3.12" and python_version < "4.0"
pyparsing==3.1.4 ; python_version >= "3.12" and python_version < "4.0"
python-dotenv==1.0.1 ; python_version >= "3.12" and python_version < "4.0"
pytube==15.0.0 ; python_version >= "3.12" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.12" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.12" and python_version < "4"
sniffio==1.3.1 ; python_version >= "3.12" and python_version < "4"
tqdm==4.66.5 ; python_version >= "3.12" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.12" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.12" and python_version < "4.0"
urllib3==2.2.3 ; python_version >= "3.12" and python_version < "4.0"
websockets==13.1 ; python_version >= "3.12" and python_version < "4.0"
yt-dlp==2024.10.7 ; python_version >= "3.12" and python_version < "4.0"
